create Validate util to handle settings validation functions
